+++ src/lib/i18n.js
@@ -65,6 +65,7 @@ const CATEGORY_MSG = {
   konut: "catKonut",
   ticari: "catTicari",
   arsa: "catArsa",
+  devren: "catDevren",
   diger: "catDiger",
 };
 const TYPE_MSG = {